+++ tests/RepoM.Api.Common.Tests/IO/ModuleBasedRepositoryActionProvider/Action/BrowseRepositoryV1Test.cs
@@ -7,6 +7,7 @@ using FluentAssertions;
 using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider;
 using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider.ActionDeserializers;
 using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data.Actions;
+using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider.Deserialization;
 using VerifyTests;
 using VerifyXunit;
 using Xunit;
