+++ tests/RepoM.Api.Common.Tests/IO/ModuleBasedRepositoryActionProvider/RepositorySpecificConfigurationTest.cs
@@ -24,6 +24,7 @@ using RepoM.Api.Common.IO;
 using RepoM.Api.Common.IO.ExpressionEvaluator;
 using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider;
 using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider.ActionMappers;
+using RepoM.Api.Common.IO.ModuleBasedRepositoryActionProvider.Deserialization;
 using RepoM.Api.Git;
 using RepoM.Api.IO;
 using VerifyTests;
@@ -36,7 +37,8 @@ public class RepositorySpecificConfigurationTest
 {
     private readonly IAppDataPathProvider _appDataPathProvider;
     private readonly MockFileSystem _fileSystem;
-    private readonly JsonDynamicRepositoryActionDeserializer _appsettingsDeserializer;
+    private readonly JsonDynamicRepositoryActionDeserializer _jsonAppsettingsDeserializer;
+    private readonly YamlDynamicRepositoryActionDeserializer _yamlAppsettingsDeserializer;
     private readonly EasyTestFileSettings _testFileSettings;
     private readonly VerifySettings _verifySettings;
     private readonly string _tempPath;
@@ -60,7 +62,8 @@ public class RepositorySpecificConfigurationTest
         _appDataPathProvider = A.Fake<IAppDataPathProvider>();
         A.CallTo(() => _appDataPathProvider.GetAppDataPath()).Returns(_tempPath);
 
-        _appsettingsDeserializer = DynamicRepositoryActionDeserializerFactory.Create();
+        _jsonAppsettingsDeserializer = DynamicRepositoryActionDeserializerFactory.Create();
+        _yamlAppsettingsDeserializer = new YamlDynamicRepositoryActionDeserializer(_jsonAppsettingsDeserializer);
 
         var dateTimeTimeVariableProviderOptions = new DateTimeVariableProviderOptions()
         {
@@ -146,7 +149,8 @@ public class RepositorySpecificConfigurationTest
             new RepositoryConfigurationReader(
                 _appDataPathProvider,
                 _fileSystem,
-                _appsettingsDeserializer,
+                _jsonAppsettingsDeserializer,
+                _yamlAppsettingsDeserializer,
                 _repositoryExpressionEvaluator));
 
         // act
@@ -172,7 +176,8 @@ public class RepositorySpecificConfigurationTest
             new RepositoryConfigurationReader(
                 _appDataPathProvider,
                 _fileSystem,
-                _appsettingsDeserializer,
+                _jsonAppsettingsDeserializer,
+                _yamlAppsettingsDeserializer,
                 _repositoryExpressionEvaluator));
 
         // act
@@ -198,7 +203,8 @@ public class RepositorySpecificConfigurationTest
             new RepositoryConfigurationReader(
                 _appDataPathProvider,
                 _fileSystem,
-                _appsettingsDeserializer,
+                _jsonAppsettingsDeserializer,
+                _yamlAppsettingsDeserializer,
                 _repositoryExpressionEvaluator));
 
         // act
@@ -224,7 +230,8 @@ public class RepositorySpecificConfigurationTest
             new RepositoryConfigurationReader(
                 _appDataPathProvider,
                 _fileSystem,
-                _appsettingsDeserializer,
+                _jsonAppsettingsDeserializer,
+                _yamlAppsettingsDeserializer,
                 _repositoryExpressionEvaluator));
 
         // act
